+++ package/index.js
@@ -9,7 +9,7 @@ export const createWebComponent = ({
   h,
   createApp,
   getCurrentInstance,
-  skipRemoveStylesOnUnmount
+  disableRemoveStylesOnUnmount
 }) => {
   if (!rootComponent) {
     console.warn('No root component provided. Please provide a root component to create a web component.')
@@ -49,7 +49,7 @@ export const createWebComponent = ({
       createApp,
       getCurrentInstance,
       elementName,
-      skipRemoveStylesOnUnmount
+      disableRemoveStylesOnUnmount
     })
   )
 }